use varuint for writing/reading objectid, total properties and property key in animation reset
objectId, propertykey are definitely written as varuint, not sure if properties total needs to be, but doesn't hurt to have it that way either I think.

I'm not sure if this is everything that needs to be updated. I tried to use `generate_core_runtime` to update the flutter runtime but that made a bunch of changes like unused imports all over the place...

Diffs=
07cbce551 use varuint for writing/reading objectid, total properties and property key in animation reset (#7510)

+++ include/rive/animation/animation_reset.hpp
@@ -21,8 +21,8 @@ private:
 public:
     AnimationReset();
     void writeObjectId(uint32_t objectId);
-    void writeTotalProperties(uint8_t value);
-    void writePropertyKey(uint8_t value);
+    void writeTotalProperties(uint32_t value);
+    void writePropertyKey(uint32_t value);
     void writePropertyValue(float value);
     void apply(Artboard* artboard);
     void complete();
